Correct ordering position for PeapExtensionsV2
It appears that some devices will fail to configure the profile,
claiming the ProfileXml we generate is corrupt.

The V1 schema specifies that PeapExtensionsV2 should come after
IdentityPrivacy.  We had it the other way around, which worked fine on
most platforms, but in some cases it appears the driver (?) checks more
strict, and (correctly) marks our profile as corrupt.

This commit corrects the order of the elements.