Do not require seq and version to be integer
The fields for version and seq are not used in the app,
but they are still mentioned in the code, in case it's needed later.
Typically, these fields will have a numeric value, but this is not
specified.

When deserializing the JSON, the value is set to NULL if it's missing,
but if the integer value is too large, an exception will be thrown.
Since we don't actually use the value, we don't need to check if the
value is valid, so it's safer to parse them as string.