Fix application on Windows 11 24H2 on ARM64
Not tested on Intel. Removes checking if network is in range,
because it requires Location permissions since 24H2.
This also caused an issue when filtering HS20 networks,
causing an invalid ConnectNetwork to be attempted.

This code may work less optimal if eduroam is out of range,
because it might try and wait for 8 seconds.